… formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/internal/ProtobufTaggedBase.kt
… formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/internal/ProtobufTaggedDecoder.kt
… formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/internal/ProtobufTaggedEncoder.kt
… formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/internal/ProtobufWriter.kt
… formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/internal/TaggedBase.kt
… formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/internal/Varint.kt
… formats/protobuf/commonTest/src/kotlinx/serialization/protobuf/ProtobufMissingFieldsTest.kt
… formats/protobuf/commonTest/src/kotlinx/serialization/protobuf/ProtobufPrimitiveWrappersTest.kt
… formats/protobuf/commonTest/src/kotlinx/serialization/protobuf/ProtobufPrimitivesTest.kt
… formats/protobuf/jvmTest/src/kotlinx/serialization/protobuf/ProtobufTopLevelPrimitivesCompatibilityTest.kt
… runtime/commonMain/src/kotlinx/serialization/Encoding.kt