* Support of top-level primitives 
* Further performance optimizations

On my machine:
```
Master:
Benchmark                         Mode  Cnt  Score   Error   Units
ProtoBaseline.fromBytes          thrpt    5  5.298 ± 0.227  ops/us
ProtoBaseline.fromBytesExplicit  thrpt    5  2.102 ± 0.040  ops/us
ProtoBaseline.toBytes            thrpt    5  6.430 ± 0.150  ops/us
ProtoBaseline.toBytesExplicit    thrpt    5  4.371 ± 0.204  ops/us


This branch:
Benchmark                         Mode  Cnt   Score   Error   Units
ProtoBaseline.fromBytes          thrpt    5  22.977 ± 0.626  ops/us
ProtoBaseline.fromBytesExplicit  thrpt    5  17.763 ± 0.856  ops/us
ProtoBaseline.toBytes            thrpt    5  20.502 ± 0.702  ops/us
ProtoBaseline.toBytesExplicit    thrpt    5  17.304 ± 2.093  ops/us
```

Fixes #216